답변 수정

+++ src/main/java/com/ll/medium/domain/home/home/controller/answer/AnswerRepository.java
@@ -2,5 +2,5 @@ package com.ll.medium.domain.home.home.controller.answer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerRepository extends JpaRepository<Answer, Long> {
+public interface AnswerRepository extends JpaRepository<Answer, Integer> {
 }
